+++ django_declarative_apis/machinery/tasks.py
@@ -11,6 +11,7 @@ from celery.task import task as celery_task
 import celery
 import time
 from typing import NamedTuple
+import kombu.exceptions
 
 from django.conf import settings
 from django.core.cache import cache
@@ -183,7 +184,29 @@ def schedule_future_task_runner(task_runner_args, task_runner_kwargs,
         countdown=countdown
     )
     task_runner_kwargs['correlation_id'] = _get_correlation_id()
-    future_task_runner.apply_async(task_runner_args, task_runner_kwargs, queue=queue, routing_key=routing_key, countdown=countdown+delay)
+
+    if settings.DECLARATIVE_ENDPOINT_TASKS_FORCE_SYNCHRONOUS:
+        logger.info('Processing tasks synchronously')
+        future_task_runner.apply(task_runner_args, task_runner_kwargs)
+    else:
+        MAX_ATTEMPTS = 3
+        for attempt in range(MAX_ATTEMPTS):
+            # XXX: This is an attempt to skirt around an unsolved, low repro issue somewhere in the celery/kombu/redis-py stack.
+            # Once in a while, a connection in the pool will timeout prior to a health check being called in redis-py and
+            # will result in an error being raised here. This should be removed once the issue has been sorted out.
+            # Note: This is around the use of redis-py in celery where celery's event loop is not running
+            # https://github.com/celery/kombu/issues/1019
+            try:
+                future_task_runner.apply_async(task_runner_args, task_runner_kwargs, queue=queue, routing_key=routing_key, countdown=countdown+delay)
+                return
+            except kombu.exceptions.OperationalError as err:
+                logger.warn('kombu.exceptions.OperationalError (attempt: %s)', attempt)
+                if attempt >= MAX_ATTEMPTS - 1:
+                    if settings.DECLARATIVE_ENDPOINT_TASKS_SYNCHRONOUS_FALLBACK:
+                        logger.warn('Falling back to executing task synchronously')
+                        future_task_runner.apply(task_runner_args, task_runner_kwargs)
+                        return
+                    raise err
 
 
 @celery_task(ignore_results=True,